backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis for background jobs
redis==5.0.1
rq==1.15.1

# AI/ML
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
langchain==0.0.350
langchain-community==0.0.10
sentence-transformers==2.2.2
chromadb==0.4.18

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23

# File handling
aiofiles==23.2.1
Pillow==10.1.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2

